Inline LanguageSupport::LanguageMethods
