+++ lib/cucumber/rb_support/rb_language.rb
@@ -7,6 +7,8 @@ require 'cucumber/rb_support/rb_transform'
 require 'cucumber/rb_support/snippet'
 require 'cucumber/gherkin/i18n'
 require 'multi_test'
+require 'cucumber/step_match'
+require 'cucumber/step_definition_light'
 
 module Cucumber
   module RbSupport
@@ -32,7 +34,6 @@ module Cucumber
 
     # The Ruby implementation of the programming language API.
     class RbLanguage
-      include LanguageSupport::LanguageMethods
       attr_reader :current_world,
                   :step_definitions
 
@@ -107,7 +108,66 @@ module Cucumber
         @current_world = nil
       end
 
+      def after_configuration(configuration)
+        hooks[:after_configuration].each do |hook|
+          hook.invoke('AfterConfiguration', configuration)
+        end
+      end
+
+      def execute_transforms(args)
+        args.map do |arg|
+          matching_transform = transforms.detect {|transform| transform.match(arg) }
+          matching_transform ? matching_transform.invoke(arg) : arg
+        end
+      end
+
+      def add_hook(phase, hook)
+        hooks[phase.to_sym] << hook
+        hook
+      end
+
+      def clear_hooks
+        @hooks = nil
+      end
+
+      def add_transform(transform)
+        transforms.unshift transform
+        transform
+      end
+
+      def hooks_for(phase, scenario) #:nodoc:
+        hooks[phase.to_sym].select{|hook| scenario.accept_hook?(hook)}
+      end
+
+      def unmatched_step_definitions
+        available_step_definition_hash.keys - invoked_step_definition_hash.keys
+      end
+
+      def available_step_definition(regexp_source, file_colon_line)
+        available_step_definition_hash[StepDefinitionLight.new(regexp_source, file_colon_line)] = nil
+      end
+
+      def invoked_step_definition(regexp_source, file_colon_line)
+        invoked_step_definition_hash[StepDefinitionLight.new(regexp_source, file_colon_line)] = nil
+      end
+
       private
+
+      def available_step_definition_hash
+        @available_step_definition_hash ||= {}
+      end
+
+      def invoked_step_definition_hash
+        @invoked_step_definition_hash ||= {}
+      end
+
+      def hooks
+        @hooks ||= Hash.new{|h,k| h[k] = []}
+      end
+
+      def transforms
+        @transforms ||= []
+      end
 
       def create_world
         if(@world_proc)
