+++ lib/cucumber/runtime.rb
@@ -4,7 +4,6 @@ require 'multi_json'
 require 'multi_test'
 require 'cucumber/configuration'
 require 'cucumber/load_path'
-require 'cucumber/language_support/language_methods'
 require 'cucumber/formatter/duration'
 require 'cucumber/file_specs'
 require 'cucumber/filters'
